Ft 137 refactor 2fa login into component (#85)
* done

* ignore lint dependency
