+++ frontend/src/components/twofalogin/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TwoFaLogin";
