+++ frontend/src/routes/SetUser.tsx
@@ -1,8 +1,7 @@
 import { useLocation, useNavigate } from "react-router-dom";
 import React, { useEffect, useState, useContext } from "react";
-import { AuthContext, IAuthContext, IUser } from "../context/auth";
-import Input from "../components/input/Input";
-import Button from "../components/button";
+import { AuthContext, IUser } from "../context/auth";
+import TwoFaLogin from "../components/twofalogin/TwoFaLogin";
 import FirstLogin from "../components/firstlogin/FirstLogin";
 
 export const BACKEND: string = `http://${window.location.hostname}:${4000}`;
@@ -10,51 +9,11 @@ export const BACKEND: string = `http://${window.location.hostname}:${4000}`;
 const SetUser: React.FC = () => {
   const nav = useNavigate();
   const location = useLocation();
-  const [redirect, setRedirect] = useState<boolean>(false);
-  const [twoFaCode, setTwoFaCode] = useState<string>("");
-  const [isFirstLogin, setIsFirstLogin] = useState<boolean>(true);
+  const [redirect, setRedirect] = useState(false);
   const [user, setUser] = useState<IUser>();
-  const auth = useContext<IAuthContext>(AuthContext);
   const [isLoading, setIsLoading] = useState<boolean>(true);
-
-  const handleTwoFaCodeSubmit = async (): Promise<void> => {
-    try {
-      const res: Response = await fetch(`${BACKEND}/twofa/login-2FA`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ twoFaCode }),
-        credentials: "include",
-      });
-
-      if (!res.ok) {
-        throw new Error(`HTTP error! Status: ${res.status}`);
-      }
-      const verified: boolean = await res.json();
-      if (verified && user) {
-        auth.logIn({
-          id: Number(user.id),
-          name: user.name,
-          intraname: user.intraname,
-          twoFAenabled: user.twoFAenabled,
-          profilePictureUrl: user.profilePictureUrl,
-          activeChats: user.activeChats,
-        });
-        setRedirect(true);
-      } else {
-        alert("Wrong 2FA Code");
-      }
-    } catch {
-      console.error("error");
-    }
-  };
-
-  const handleTwoFaCodeInputChange = (
-    e: React.ChangeEvent<HTMLInputElement>
-  ): void => {
-    setTwoFaCode(e.target.value);
-  };
+  const auth = useContext(AuthContext);
+  const [isFirstLogin, setIsFirstLogin] = useState<boolean>(false);
 
   useEffect(() => {
     const urlParams = new URLSearchParams(location.search);
@@ -80,8 +39,8 @@ const SetUser: React.FC = () => {
           }
         })
         .catch((err) => console.error(err));
-      setIsLoading(false);
     }
+    setIsLoading(false);
   }, [auth]); // eslint-disable-line react-hooks/exhaustive-deps
 
   useEffect(() => {
@@ -91,25 +50,13 @@ const SetUser: React.FC = () => {
   }, [redirect]); // eslint-disable-line react-hooks/exhaustive-deps
 
   if (isLoading) {
-    return <div></div>;
+    return <div>Loading</div>;
   }
 
   return (
     <>
       {user?.twoFAenabled && (
-        <div>
-          <div>
-            <Input
-              label="2FA code"
-              placeholder="Enter 2FA code here"
-              value={twoFaCode}
-              onChange={handleTwoFaCodeInputChange}
-            ></Input>{" "}
-          </div>
-          <div>
-            <Button onClick={handleTwoFaCodeSubmit}>Submit 2FA Code</Button>
-          </div>
-        </div>
+        <TwoFaLogin setRedirect={setRedirect} user={user} />
       )}
       {isFirstLogin && <FirstLogin />}
     </>
